More work on grapple
So before, you could glitch the grapple by flicking your view to have the guntip be far from the grappleLocation and cause problems. NOW, the gun rotates to face the grappleLocation so gunTip cant be flicked and will always finish the while loop, then the gun will loop again to put it back to normal orientation. HOWEVER, high speeds have a chance to glitch the grapple, but that can be done later.